Fixed compilation of QtV8 for mac
configure is currently whitelisting a set of architectures which are
known to be capable of building V8.  macosx was omitted from this list.

Note that we don't attempt to differentiate between different macosx
architectures; we expect any mac which is capable of building Qt5 at all
is also capable of building V8.

Change-Id: I345f2a446d650cbeb1e2ad81656cdbeea6fa0d9a
Reviewed-on: http://codereview.qt.nokia.com/4293
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>