Fix for wayland config.test
Wayland config.test would only be able to build if there was a valid
value for QMAKE_LIBDIR_WAYLAND because if this was empty, -L would be
used, which would lead to invalid command syntax.  So even if the
library was already in the build path, the test would fail.  The test
now checks QMAKE_LIBDIR_WAYLAND for valid directoires first before add
them to the LIBS variable.

Change-Id: Ic3a3be54a41390d6302a122a6d20f896e962ccc7
Reviewed-on: http://codereview.qt.nokia.com/4354
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Paul Olav Tvete <paul.tvete@nokia.com>