Ensure that the proxystyle is used over the style's own functions
When calling a style function such as pixelMetric() then this should be
done via the proxy style and not on the style directly.  This is so that
the proxy style always has a chance to override the original style's
implementation and still preserve the fallback to the base style.

Task-number: QTBUG-20849
Merge-request: 2645
Reviewed-by: olivier

(cherry picked from commit 270c20490813fa943b4fbf0feb1aecb98f0b19ba)

Change-Id: Id43ec17fd1eae0136d6e1f050f23bca6d7d6d6cf
Reviewed-on: http://codereview.qt-project.org/4591
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>