Add some enablers to QTextDocumentLayout
Required to access internals used in the layout process, so that
we can present the layout in the scenegraph. For instance, the
frameBoundingRect() for tables will be the internal bounding rect
of the cells. To avoid regressions, I've added a separate function
for the outer bounding rect. For now, this is private API, but if
it turns out to be useful, the functions can easily be added as
virtual to QAbstractTextDocumentLayout.

Task-number: QTBUG-20917
Change-Id: I04e07d3850b21f9f0704bf8c8a3ffe97ee5c3fda
Reviewed-on: http://codereview.qt-project.org/5539
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Jiang Jiang <jiang.jiang@nokia.com>