Don't build qitemview autotest on IRIX.
The test was being skipped one test function at a time on IRIX.  It's
better to avoid building or running it, rather than spending time
running a test that does nothing on that platform.

This commit also turns a silent skip on other platforms into an explicit
QSKIP to make it more obvious that there's an unfixed TODO there.

Change-Id: I7cae4f0f2eaebf5977a781b2ce15d1e71ef13b36
Reviewed-on: http://codereview.qt-project.org/5957
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>