Avoid building empty, inapplicable QWS tests.
It is better to avoid building a test when we know that it is not
applicable, rather than building an empty test that appears to pass.
Note that some of the tests were already excluded by
tests/auto/other.pro, but still had the preprocessor directives to make
an empty test.

Change-Id: I71b75c37a05f6e20dbef43d82f72b29b35ef496a
Reviewed-on: http://codereview.qt-project.org/6410
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>