Make CONFIG+=GNUmake respect shadow builds
Previously we were generating include lines based on the relative path of
the source file, which resulted in the .d files being placed in the source
dir as well. We now expect the .d files to live in the output dir, but keep
the dependency from the .d file to the original source file.

Before:

   .deps/%.d: %.cpp
   -include .deps/../../src/foo.d

After:

   .deps/%.d: ../../src/%.cpp
   -include .deps/foo.d

Change-Id: I749adeb671cf8424f0849521c5bb1489eb3e76d5
Reviewed-on: http://codereview.qt-project.org/6455
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Marius Storm-Olsen <marius.storm-olsen@nokia.com>